task_manager/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
python-decouple==3.8